added uh guard for file deletion

+++ laravel/app/Http/Controllers/MediaController.php
@@ -4,7 +4,9 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\MediaRequest;
 use App\Http\Resources\MediaResource;
+use App\Models\Blog;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Spatie\MediaLibrary\MediaCollections\Models\Media;
 use Spatie\QueryBuilder\AllowedFilter;
@@ -20,10 +22,6 @@ class MediaController extends Controller
 
     public function index(): AnonymousResourceCollection
     {
-        // TODO: pagination per media type (image/audio/video)
-        // - pass type param to indexMediaApi
-        // - filter in Laravel with mime_type like
-        // - cache fetched state per type in store
         $media = QueryBuilder::for(Media::class)
             ->allowedFilters([
                 AllowedFilter::partial('mime_type'),
@@ -41,13 +39,27 @@ class MediaController extends Controller
         return new MediaResource($media);
     }
 
-    public function destroy(int $id): JsonResponse
+    public function destroy(int $id, Request $request): JsonResponse
     {
         /*
            * another way to bind model from 3rd party package using this
            * Route::model('media', \Spatie\MediaLibrary\MediaCollections\Models\Media::class);
            */
         $media = Media::findOrFail($id);
+
+        if (! $request->boolean('force')) {
+            $inUse = Media::where('model_type', Blog::class)
+                ->where('original_url', $media->original_url)
+                ->exists()
+                || Blog::where('thumbnail', $media->original_url)->exists();
+
+            if ($inUse) {
+                return response()->json([
+                    'message' => 'Media is in use. Are you sure you want to delete it?',
+                    'in_use' => true,
+                ], 422);
+            }
+        }
         $media->delete();
 
         return response()->json([
